+++ apps/tempest.games/src/backend.bun.ts
@@ -13,8 +13,8 @@ import { createServer } from "../dev/https-dev"
 import { worker } from "./backend.worker"
 import { db } from "./backend/db"
 import { logger, parentSocket } from "./backend/logger"
-import type { Context, ContextAuth } from "./backend/router"
 import { appRouter } from "./backend/router"
+import type { Context, ContextAuth } from "./backend/trpc-server"
 import { userSessions } from "./backend/user-sessions"
 import { serveSocket, sessionMiddleware } from "./backend/websockets"
 import { env } from "./library/env"
